Css Cafe Menu
Build a Cafe Menu using Css (Selectors, Display properties, Background Image,  ) and Html (Img, Article, Section, Class, Link, ). It took inspiration form FreeCodeCamp to build this project.

Learning
I learnt about difference between inline and block level element, how to center block elements, how to give different colors on a link text for eg: when someone click on it or hover on it or when the link is clicked then which color should remain. How to build a layout and how to structure a html accordingly.